fixed few bugs in status cmd

+++ src/taskmaster/process/kill_processes.c
@@ -8,6 +8,8 @@ void	kill_processes(int signum, const t_lst_proc *procs)
 	for (; procs; LTONEXT(procs))
 	{
 		proc = LCONT(procs, t_process*);
+		if (proc->pid <= 1)
+			continue;
 		if (kill(proc->pid, signum) == -1)
 		{
 			if (errno != ESRCH)
